+++ Review+CoreDataProperties.swift
@@ -19,7 +19,7 @@ extension Review {
 
     @NSManaged public var desc: String?
     @NSManaged public var rating: Int16
-    @NSManaged public var reviewID: Int32
+    @NSManaged public var reviewID: Int64
     @NSManaged public var reviewed: User?
     @NSManaged public var reviewer: User?
